Merging with original 1.13.x branch

+++ src/main/java/moze_intel/projecte/gameObjs/items/tools/PEToolBase.java
@@ -92,7 +92,7 @@ public abstract class PEToolBase extends ItemMode
 	/**
 	 * Clears the given tag name in an AOE. Charge affects the AOE. Optional per-block EMC cost.
 	 */
-	protected void clearTagAOE(World world, ItemStack stack, EntityPlayer player, Tag<Block> tag, int emcCost, EnumHand hand)
+	protected void clearTagAOE(World world, ItemStack stack, EntityPlayer player, Tag<Block> tag, long emcCost, EnumHand hand)
 	{
 		int charge = getCharge(stack);
 		if (charge == 0 || world.isRemote || ProjectEConfig.items.disableAllRadiusMining.get())
@@ -133,7 +133,7 @@ public abstract class PEToolBase extends ItemMode
 	/**
 	 * Tills in an AOE. Charge affects the AOE. Optional per-block EMC cost.
 	 */
-	protected void tillAOE(ItemStack stack, EntityPlayer player, World world, BlockPos pos, EnumFacing sidehit, int emcCost)
+	protected void tillAOE(ItemStack stack, EntityPlayer player, World world, BlockPos pos, EnumFacing sidehit, long emcCost)
 	{
 		int charge = this.getCharge(stack);
 		boolean hasAction = false;
@@ -272,7 +272,7 @@ public abstract class PEToolBase extends ItemMode
 	/**
 	 * Carves in an AOE. Charge affects the breadth and/or depth of the AOE. Optional per-block EMC cost.
 	 */
-	protected void digAOE(ItemStack stack, World world, EntityPlayer player, boolean affectDepth, int emcCost, EnumHand hand)
+	protected void digAOE(ItemStack stack, World world, EntityPlayer player, boolean affectDepth, long emcCost, EnumHand hand)
 	{
 		if (world.isRemote || this.getCharge(stack) == 0 || ProjectEConfig.items.disableAllRadiusMining.get())
 		{
@@ -344,7 +344,7 @@ public abstract class PEToolBase extends ItemMode
 	/**
 	 * Attacks in an AOE. Charge affects AOE, not damage (intentional). Optional per-entity EMC cost.
 	 */
-	protected void attackAOE(ItemStack stack, EntityPlayer player, boolean slayAll, float damage, int emcCost, EnumHand hand)
+	protected void attackAOE(ItemStack stack, EntityPlayer player, boolean slayAll, float damage, long emcCost, EnumHand hand)
 	{
 		if (player.getEntityWorld().isRemote)
 		{
@@ -405,7 +405,7 @@ public abstract class PEToolBase extends ItemMode
 	/**
 	 * Shears entities in an AOE. Charge affects AOE. Optional per-entity EMC cost.
 	 */
-	protected void shearEntityAOE(ItemStack stack, EntityPlayer player, int emcCost, EnumHand hand)
+	protected void shearEntityAOE(ItemStack stack, EntityPlayer player, long emcCost, EnumHand hand)
 	{
 		World world = player.getEntityWorld();
 		if (!world.isRemote)
